at last, refactor old database and DB-related code

+++ MusCat.Core/Entities/Lineup.cs
@@ -2,8 +2,8 @@ namespace MusCat.Core.Entities
 {
     public class Lineup
     {
-        public long PerformerId { get; set; }
-        public long MusicianId { get; set; }
+        public int PerformerId { get; set; }
+        public int MusicianId { get; set; }
         public short? YearStart { get; set; }
         public short? YearEnd { get; set; }
         public Musician Musician { get; set; }
