+++ MusCat.Core/Entities/Song.cs
@@ -5,8 +5,8 @@ namespace MusCat.Core.Entities
 {
     public class Song : IDataErrorInfo
     {
-        public long Id { get; set; }
-        public long AlbumId { get; set; }
+        public int Id { get; set; }
+        public int AlbumId { get; set; }
         public byte TrackNo { get; set; }
         public string Name { get; set; }
         public string TimeLength { get; set; }
