+++ MusCat.Core/Interfaces/Domain/IAlbumService.cs
@@ -7,10 +7,10 @@ namespace MusCat.Core.Interfaces.Domain
     public interface IAlbumService
     {
         Task<Result<Album>> AddAlbumAsync(Album album);
-        Task<Result<Album>> RemoveAlbumAsync(long albumId);
+        Task<Result<Album>> RemoveAlbumAsync(int albumId);
         Task<Result<Album>> UpdateAlbumAsync(Album album);
-        Task<Result<Album>> UpdateAlbumRateAsync(long albumId, byte? rate);
-        Task<Result<Album>> MoveAlbumToPerformerAsync(long albumId, long performerId);
-        Task<IEnumerable<Song>> LoadAlbumSongsAsync(long albumId);
+        Task<Result<Album>> UpdateAlbumRateAsync(int albumId, byte? rate);
+        Task<Result<Album>> MoveAlbumToPerformerAsync(int albumId, int performerId);
+        Task<IEnumerable<Song>> LoadAlbumSongsAsync(int albumId);
     }
 }
