+++ MusCat.Core/Interfaces/Domain/ICountryService.cs
@@ -5,9 +5,9 @@ namespace MusCat.Core.Interfaces.Domain
 {
     public interface ICountryService
     {
-        Task<long> GetPerformersCountAsync(byte countryId);
+        Task<int> GetPerformersCountAsync(int countryId);
         Task<Result<Country>> AddCountryAsync(string name);
-        Task<Result<Country>> RemoveCountryAsync(byte countryId);
-        Task<Result<Country>> UpdateCountryAsync(byte countryId, string name);
+        Task<Result<Country>> RemoveCountryAsync(int countryId);
+        Task<Result<Country>> UpdateCountryAsync(int countryId, string name);
     }
 }
