+++ MusCat.Core/Interfaces/Domain/IPerformerService.cs
@@ -6,10 +6,10 @@ namespace MusCat.Core.Interfaces.Domain
     public interface IPerformerService
     {
         Task<Result<Performer>> AddPerformerAsync(Performer performer);
-        Task<Result<Performer>> RemovePerformerAsync(long performerId);
+        Task<Result<Performer>> RemovePerformerAsync(int performerId);
         Task<Result<Performer>> UpdatePerformerAsync(Performer performer);
-        Task<Result<Album>> AddAlbumAsync(long performerId, Album album);
-        Task<Result<Country>> GetCountryAsync(long performerId);
-        Task<long> SongCountAsync(long performerId);
+        Task<Result<Album>> AddAlbumAsync(int performerId, Album album);
+        Task<Result<Country>> GetCountryAsync(int performerId);
+        Task<int> SongCountAsync(int performerId);
     }
 }
