+++ MusCat.Core/Interfaces/Domain/ISongService.cs
@@ -6,7 +6,7 @@ namespace MusCat.Core.Interfaces.Domain
     public interface ISongService
     {
         Task<Result<Song>> AddSongAsync(Song Song);
-        Task<Result<Song>> RemoveSongAsync(long SongId);
+        Task<Result<Song>> RemoveSongAsync(int SongId);
         Task<Result<Song>> UpdateSongAsync(Song newSong);
     }
 }
